Generic system to identify sources of resources
Food sources can now be identified by a social unit and this has been implemented in a way so we can extend it for other resources or other ways to gather the food.

The Social Unit doesn't need to understand how to harvest grain or how to hunt a rabbit, the food source will handle that.